fix : edit_job  progress_status + progress_remark/rem confusion: 1. the progress_status  was not working/updating because ajax was not passing values through.
2. progress_remark / rem confusion:

progress_remark is used to describe and annotate the progress of the scene/rendering, for instance "need to fix spot buffer shadows" or need to wait until John has finished animating the dragon". it apperas in the job list under "progress status" column.

rem is now only used as a generic remark field for the job in general, it will now only appear under title on view_job page.